--- src/core/assistant-message/directives/index.ts
@@ -1,5 +1,0 @@
-export { StreamingParser } from "./StreamingParser"
-export { TextContentHandler } from "./TextContentHandler"
-export { ToolUseHandler } from "./ToolUseHandler"
-export { ParameterHandler } from "./ParameterHandler"
-export type { TextDirective, ToolDirective, Directive, ParsingState } from "./types"
