+++ src/core/assistant-message/parsers/types.ts
